requirements.txt
google-generativeai>=0.3.1
openai>=1.0.0
anthropic>=0.5.0
requests>=2.28.0
grpcio==1.60.1